+++ contracts/mocks/IexecPocoMock.sol
@@ -1,0 +1,30 @@
+// SPDX-FileCopyrightText: 2024 IEXEC BLOCKCHAIN TECH <contact@iex.ec>
+// SPDX-License-Identifier: Apache-2.0
+
+pragma solidity ^0.8.20;
+
+import {IexecLibOrders_v5} from "@iexec/poco/contracts/libs/IexecLibOrders_v5.sol";
+import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
+
+/**
+ * @notice Testing purposes only.
+ */
+contract IexecPocoMock is ERC20 {
+    constructor() ERC20("Staked RLC", "SRLC") {
+        _mint(msg.sender, 1000000);
+    }
+
+    function sponsorMatchOrders(
+        IexecLibOrders_v5.AppOrder calldata appOrder,
+        IexecLibOrders_v5.DatasetOrder calldata datasetOrder,
+        IexecLibOrders_v5.WorkerpoolOrder calldata workerpoolOrder,
+        IexecLibOrders_v5.RequestOrder calldata requestOrder
+    ) external returns (bytes32 dealId) {
+        requestOrder; // Silent warning
+        uint256 dealPrice = appOrder.appprice +
+            datasetOrder.datasetprice +
+            workerpoolOrder.workerpoolprice;
+        _burn(msg.sender, dealPrice);
+        return keccak256("deal");
+    }
+}